Refactor and clean up InstanceOfD/Verify*Type irgen
Summary: These two functions want almost exactly the same thing when the input
is an object, but instead of calling a common helper they were a mishmash of
copypasta, with an important interface case missing from verifyTypeImpl(). This
diff pulls most of the logic into a common helper, cleaning it up a bit along
the way.

Reviewed By: @jdelong

Differential Revision: D2144982