Compute vtables for interfaces, use for method dispatch and instance checks
Summary: hhbbc assigns every interface a vtable slot, with the guarantee that
no two interfaces implemented by the same class will share the same slot. Then,
when constructing a Class, we build a table of vtables, with one entry for
every interface it implements. The vtable table contains both vtable pointers
as well as the Class* itself. The vtable pointers are used for method dispatch
when the jit knows that the base object implements a given interface, and the
Class* are used for efficient instance checking, using the same pattern as
normal class instance checks and m_classVec.

Reviewed By: @markw65

Differential Revision: D2145607