Provide option to disable Zend Ini compatibility
Summary: D1797805 enabled us to have better compatibility between ini and our
zend compatibility layer, but it broke certain per request type
of settings like `upload_max_filesize`. You could `ini_get()` these
values running a CLI script, but server requests gave you a `bool(false)`
signifying that the settings were not being bound.

This is a hotfix for our LTS releases so that we can keep the current
functionality but unbreak people as well. Basically provide a
rutnime option to disable the zend ini compatibility.

Longer term, something around D2099778 will be the anwser to the
problem. But we need to let that review and bake in `master` for
a while.

Ref. #4993

Reviewed By: @jwatzman

Differential Revision: D2148517