hash-crc* mis-predicted branch fix
The crc32 unaligned version incurs a penalty when using the loop
instruction since there is a swtich between the LSB and MITE de-
coder. This will ensure that the following sequence of

dec     %ecx
jmp/jcc LABEL

micro-ops get fused without the added overhead of switching
decoders.

The branch misprediction issue still continues to exist inspite of
the conditional jump so that's still a WIP.