Add some splitting heuristics for intervals
Summary: The new CallBuiltin code in D2149221 results in
some really strange register allocation.

Inlined NativeImpls can end up with an EagerSyncPoint for each
argument, and the "pc" in each one is the same (the address of
a bytecode instruction). The result is that xls keeps it live
across a lot of blocks that don't care about it; and the only
blocks that /do/ care, are side-exits of some kind. Worse, it
ends up materializing it on multiple edges into the same blcok,
which then result in artifical (and bizarre) control flow of
the form:

      moveq #0xdeadbeefcafe, $r14
      jmp B10
  B11 moveq #0xdeadbeefcafe, $r14
  B10 ... // nothing with $r14
      ...
  B54<unlikely> store $r14

These two heuristics fix the problem, and it materializes the pc
in each of the exit blocks, at the point of use.

I'm pretty sure the first one is never bad. The second one might
cause non-constants to be spilled when they didn't need to be,
so I restricted it to just applying to constants.

Reviewed By: @edwinsmith

Differential Revision: D2154387