shave down one size class
Summary: Reduce kMaxSmartSize to 7/8 of kSlabSize, so that a DebugHeader
always fits. Also, whenever we needed a "smart" block of the max
size, we were pretty much guaranteed to fragment the current slab,
*and* fill the next slab.

This also eliminates the annoying fact that in debug builds,
kMaxSmartSize is not an even size class, due to the padding for
a DebugHeader. (I've debugged problems from that several times).

Reviewed By: @binliu19

Differential Revision: D2151700