Preallocate sf flags in xls
Summary: Vasm-xls, our PhysReg encoding, x64, and AArch64 all only support
one live status-flags register at a time. We currently allocate
SF registers the usual (fully general) way, but also have to compute
the union of all SF's so we can safely insert xors for the ldimm{0}
optimization.

This diff adds a checkSF() to enforce the 1-SF register rule in vasm,
and modifies vasm-xls to eagerly rename VregSF to RegSF{0} without
computing intervals for each one. This reduces the number of intervals
we need to process and eliminates the sf-union calculation.

Reviewed By: @markw65

Differential Revision: D2155548