Make ObjectData ref-counts start at 1
Summary: Similar to how StringData worked, the initial ref-count for ObjectData
started at zero, but was always eventually incremented to one before being
used. This is inefficient, so instead have the ref-count for ObjectData start at
one, and remove the now redundant inc-refs.

This causes a fair amount of code churn, but its mainly changing calls to the
Object constructor (which does an inc-ref), to use Object::attach instead. This
diff should strictly remove instructions.

Reviewed By: @edwinsmith

Differential Revision: D2140478