Fix alias_torture reliance on a particular RNG
Summary: Although this test picks a specific fixed seed, it's still
reliant on whatever the libc implementation of `rand` is in order to get
consistent results. This means that the test fails on OS X for no good
reason, since the libc there has a different RNG. Instead, just use the
`mt_rand` family of functions, which use the same RNG across platforms.

Reviewed By: @paulbiss

Differential Revision: D2160574