Cleanup some runtime->ext dependencies
Summary: Ideally we don't want the core runtime to be
calling into extension code. This removes some potentially
expensive wrapper code (eg constructing a Variant, just to
pass a reference to it into is_callable, which will be
ignored because isReferenced will always be false).

There could be a small overhead in this version, because
the HHVM_FE(is_callable) now forwards to the is_callable. After
an upcoming diff, php will just use the runtime version.

Reviewed By: @swtaarrs

Differential Revision: D2164210