Structural subtyping for shapes proposal
Summary: This makes

  type S2 = shape(
    'x' => int,
    'y' => string,
  );

subtype of

  type S1 = shape(
    'x' => int,
  );

which makes sense from [[ http://en.wikipedia.org/wiki/Liskov_substitution_principle | LSP ]] point of view - the only thing you can do about S1 is set or get its 'x' index, which is an int, and exactly the same can be said about S2, so S2 can be a subtype of S1. This in itself is kind-of useful, but is also a prerequisite to another step, which is anonymous shapes as typehints. This is motivated by trying to encourage people to use (better typed) shapes where they would use arrays:

  function f(array<string, int> $v): int {
    return $v['x']*2;
  }

could be even better typed as

  function f(shape('x' => int)  $v): int {
    return $v['x']*2
  }

The drawback of this approach is illustrated by the test case that used to fail but is now passing (diffusion/FBCODE/browse/master/hphp/hack/test/typecheck/shape3.php):

  type myshape = shape(
    'field1' => int,
    'field2' => bool,
  );

  function test(): myshape {
    $x = shape('field1' => 1, 'field2' => true, 'field3' => 'test');
    return $x;
  }

Some fields might be set unnecessarily, and we won't catch it in the new model. To make up for it, {D1925231} guarantees that all fields are set, so the only case where it makes us more blind are:

- doing unnecessary work (i.e forgetting to remove something after refactoring)
- setting field repeatedly and misspelling it:

  type S = shape (
    'x' => int,
  );

  function f(): S {
    $s = shape();
    $s['x'] = 3;
    ...
    $s['x']++;
     ...
    $s['y'] ++; // oops, meant 'x' here
  }

Reviewed By: @dlreeves

Differential Revision: D2019698