uniqify filename of eval()'d code
ParserBase::newClosureName determine the name of a Closure based
on the filename, a per-parse counter, and some other information.
When Closure's are generated from inside eval()'d code and the
same eval() statement is used  all of this information is the same
and can result in the same Closure name being used multiple times
causing unexpected behavior.

This patch adds an md5 of the code being eval'd to what is considered
the filename of this eval'd code to uniqify different bits of code
eval()'d from the same line.