Fix cloning of pseudomain Funcs
Summary: D2117197 introduced an error where the first time we clone a pseudomain,
we scribble over its Class contexts and alias it in a hashmap.  This
leads to a double free when we destroy the Unit (which, notably, only
happens in non-repo mode), and probably causes incorrect behavior when
executing code included or required from class contexts.

Reviewed By: @markw65, @paulbiss

Differential Revision: D2167085