Use region predicted type support to fix inner-ref guard relaxation problem
Summary: If a guard is relaxed away to Gen, but we had a inner ref prediction,
keep it in the region prediction list.  Make translate-region emit predictions
if they are there (before this diff they never are there without guard
constriction on, or due to a bug in region-method), so we'll get it when we
translate later.

One odd thing for now, is this does mean that when the tracelet is selected, we
see a HintLocInner or HintStkInner to get inner-ref types, and then when it's
translated that can turn into a PredictLoc or PredictStk.  We probably can
remove the Hint instructions now though---and it looks like the difference is
harmless here (it looks like they assert the thing is boxed and then handle
predicting TBottoms differently).

Reviewed By: @swtaarrs

Differential Revision: D2169224