Clone Classes instead of Funcs when scoping closures
Summary: Currently, when we scope a closure object, we make a clone of its
function body (i.e., its __invoke() method) and chain it in a linked
list of differently-scoped invoke methods.  This results in a lot of
special casing across the codebase to handle calls to closures---in
particular, all calls to a given closure would hit the pre-scoped
__invoke, which necessitated weird hijinks in func prologues to dispatch
to the appropriately scoped body.

In repo auth mode, when traits are flattened and dynamic closure scoping
is banned, a particular closure is only ever scoped at the time of its
creation.  Since each closure is given a unique subclass of Closure to
instantiate from, we can simply reuse this Class and its __invoke() for
the first scoping---this lets us make closure calls look like all other
calls, and in particular lets us avoid an indirect jump to the closure's
DV init at the end of every closure prologue.

When a closure /does/ need to be re-scoped after the first scoping, we
make a clone of the Closure subclass instead of just cloning the
__invoke().  This uses more memory than a clone of a Func, but saves
dispatch complexity.  Moreover, in repo mode, we are now cloning one
fewer Func per closure, since we simply reuse the existing "pre-scope"
Class and Func for the first scoping event.

Because we need to set the instance bits correctly on the closure
subclasses, we also rewrite the AllClasses iterable using interfaces
compatible with range-based-for and refactor the locking around
InstanceBits initialization.

This change has the side-effect of fixing reflection on non-scalar
default values of parameters in closures defined within namespaces.

Reviewed By: @jdelong

Differential Revision: D2133253