Disable slow/ext_string/setlocale-threaded.php in repo mode
Summary: Seems to pass in contbuild, but this test has been failing locally for
me for months, apparently only in repo mode.  It's pretty clear the problem is
that the test is too sensitive to environment, so just disable it.  (It looks
like it's doing a bunch of path stuff that it probably shouldn't do, and the
way it fails for me in repo mode is saying it can't find its file.)

Reviewed By: @alexmalyshev

Differential Revision: D2169267