Don't always append a second ini file to the test command
Summary: Right now, for every test with an explicit ini file associated with it, we append a second ini file to the command in `/tmp`

```
../_bin/hphp/hhvm/hhvm -c test/slow/program_functions/ini_binding.php.ini -vEval.EnableArgsInBacktraces=true -vEval.EnableIntrinsicsExtension=true -vRepo.Local.Mode=-- -vRepo.Central.Path=/data/users/joelm/fbcode-baseline/_bin/hphp/hhvm/verify.hhbc -vEval.Jit=true      -vResourceLimit.CoreFileSize=0    --file 'test/slow/program_functions/ini_binding.php' -c /tmp/ini_binding.phpCLWgn9.ini
```

This is uncessarily duplicatitve for nearly every test. There are just a handful of
tests that require it (e.g. some `auto_prepend` tests), and they have the marker
to show they require it (i.e., `{PWD}`). So let's just append this extra ini
file for those tests

Reviewed By: @jwatzman

Differential Revision: D2165974