Standard library for shapes - runtime support
Summary: Shapes are integral part of the language, but they are missing even basic support in standard library - in runtime, the array functions work with them (since they are arrays under the hood), but if those array functions are typed in .hhi files (as taking array, KeyedContainer, Indexish, etc.) the typechecker will balk because there is no implicit subtyping relation between shapes and arrays (by design).

They also can't be implemented in user space, because we are missing a way to express things like "this function accesses shape field" using type hints.

We could invent a completely new notations to express it, but since shapes are relatively simple features - you can basically only set, access or unset a field - we decided to start by hardcoding a handful of those functions in the typechecker in the form of Shapes class, mimicking helper functions in Vectors, Sets etc. - see {D2116812} for example. I make it auto imported in hh files because we plan to consider them a part of the language now. In particular, {D2094934} introduces an error message forbidding accessing optional fields in any other way than using functions from this class.

In runtime they are implemented as simple wrappers around array functions.

Reviewed By: @jwatzman

Differential Revision: D2154441