Fix xdebug with multiple commands (and hopefully with PhpStorm)
Summary: Thanks to lethalman for finding the problem on
https://github.com/facebook/hhvm/issues/4488 -- if a `recv` managed to
read multiple commands at once, we'd use the first one, and then drop
the rest on the floor. Instead, we move through the buffer, consuming as
we go, and refilling when needed. This isn't super pretty -- IMO there's
too much object state going on here -- but I think is the best we can do
with the current code.

Fixes https://github.com/facebook/hhvm/issues/4488

Reviewed By: @alexmalyshev

Differential Revision: D2170147