A bit of preg cleanup
Summary: This is to facilitate the FCallBuiltin streamlining work.
When a builtin is rewritten to use the faster calling conventions,
its no longer callable from c++. The preg functions already have
better alternatives that should be used anyway, so this diff
just updates everything to use them, and tweaks the api slightly
to suit the new calling conventions at the same time.

Reviewed By: @swtaarrs

Differential Revision: D2175100