Fix guard constriction issues
Summary: If the first instruction in a region was interpOned or isAlwaysNop(),
we were leaving translateInstr() before emitting an ExitPlaceholder. This was
causing visitGuards() to walk the whole IRUnit, visiting different Check*
instructions for the same location and deciding that the type to guard was
Bottom.

Also get rid of some unnecessary lambdas in emitPredictionsAndPreconditions()
and use DataTypeGeneric for all inputs to NewPackedArray, to be consistent with
its dontGuard input attribute. This also prevents some pathological behavior in
a couple zend tests that give >1000 args to min/max.

Reviewed By: @jdelong

Differential Revision: D2176092