Save an instruction in the CRC hash functions
Summary: Apply a trick used in D1765309 to save an instruction in the aligned version of
the CRC hash functions.

It shouldn't matter much to performance, but should allow the unaligned version
to include the aligned versions while still fit the frequent path within a
cache line, which helps perf on non-Facebook workload. @

Reviewed By: @markw65

Differential Revision: D2176909