Don't emit unreachable jump in irgen::endBlock, and make irgen::beginInlining more robust
Summary: irgen::endBlock was inserting fall-through jumps even when they're
unreachable.  This would save a FrameState for the target block, thus
causing it to be processed by the region translator even though it's
unreachable, which would trigger an assert in irgen::beginInlining if
the FPush* gets a TBottom and thus punts.  So avoid this by only
emitting the fall-through jump in endBlock if the jump is reachable.

This diff also makes irgen::beginInlining more robust in the situation
described above.  Instead of failing an assert, it now bails and
translateRegion then skips to the next reachable block.

Reviewed By: @jdelong

Differential Revision: D2173873