resolve symlinks when looking up package.json
Summary: instead of following symlinks when looking for all `package.json` files (we had issues with `find` following cycles), only find `package.json` files whose hardlinks are within the root dir, and resolve the symlink when someone calls `require` on a symlinked module.

that is, if you do `require('foo')` inside `/path/to/root/subdir/file.js`, we'll look for `/path/to/root/subdir/node_modules/foo/package.json`, then `/path/to/root/node_modules/foo/package.json`, then `/path/to/node_modules/foo/package.json`, etc. if one of those paths is a symlink, we'll follow it. this must point at a module inside the root dir, so that our prescan will have found it.

in the future, we should just parse the module if the symlink points outside of the root dir, but this gets us basic symlink support without erroring.

Reviewed By: @gabelevi

Differential Revision: D2178728