Make sure iteration of class_iterator actually stops
Summary: The comparison here in `.equal` isn't quite right -- if, in
`next_bound_name_if_null`, we null out `m_cls`, then `m_clone` might
point to the end of some random iterator -- and notably not the same as
`all_classes().end()`. On the version of Boost we use, I suspect all end
iterators are the same, but with Boost 1.58 on my Mac, they are not.
This means that callers won't actually stop iterating, since we won't
compare equal to the end iterator, will iterate off the end of the
iterator, and (thankfully) trip the helpful assert that `m_cls` isn't
null.

Reviewed By: @mxw

Differential Revision: D2180716