Refactor irgen::emitFuncPrologue()
Summary: Magic calls require splitting the prologue into two parts, the "prologue
entry", which ends after the early stack check, and the "prologue body",
which is everything else, because magic call prologues have is-magic
checks between the two parts (and sometimes have two prologue bodies).

Let's split the function up in a separate diff to make the magic call
conversion a bit cleaner.  We also refactor the stack overflow check
logic, since it's used in both halves of the prologue.

Reviewed By: @ottoni

Differential Revision: D2174608