Memory cleanup in hphp/compiler
Summary: About 10% of total memory was in LocationPtrs.
Each LocationPtr contains 4 ints that we use for source
positions, and char* saying which file it came from, and
an int cursor which doesn't seem to be used much.

Every Construct contains a LocationPtr, which is 16
bytes, the same as the 4 ints. Each Construct can also
get a pointer to its FileScope, which has the file name,
so we can get away with just storing the 4 ints on each
Construct. The result is no increase in size for Construct,
and we get rid of all the memory that was being used for
the Locations.

The tricky part is that during parsing, the scope-tree
hasn't been created yet, so there's no way to get from a
Construct to its FileScope - so a FileScopeRawPtr had to
be plumbed through to various places that previously
didn't need it.

Reviewed By: @jdelong

Differential Revision: D2191233