Revert "Added SimpleXMLIterator"
Summary: This reverts commit a117c525f038a25673762ad89e0e3f8a30d9c461.

It is causing an internal test to fatal that passes on the previous release.
It looks like it's related to how this change is now attempting to call a
getIteratorForTraversable method, which the mocking framework gets confused
about---generally, the diff is making a user-observable change to how object
iteration works.  (And it changes whether some scripts fatal depending on the
return value of a getIterator method on an IteratorAggregate (e.g. P19893893).)
I'm not sure what the best solution is, but it might be more reasonable (as in
more observably-compatible with php behavior) to just hard code
c_SimpleXMLElement::classof here?  That class is so special in general that it
seems like the usual way to handle it ...

Reverts #5352

Reviewed By: @alexmalyshev

Differential Revision: D2193110