CGetPropQ should unbox refs
Summary: CGetPropQ should always produce a cell, but it wasn't properly checking
for refs and unboxing them. This causes incorrect behavior when using the
null-safe access operator on a method which returns a ref. With a debug build it
would typically cause an assertion to trigger when the ref was encountered when
a cell was expected. Add the logic to unbox as necessary.

This was actually fixed as part of commit
230fe2e641b8a537ac4983a4a5d44539bfbac041, but that was reverted in
cc58e45897728a63811f47c1673ff47e42ad688c.

Also add a new test case which exercises this behavior.

Reviewed By: @alexmalyshev

Differential Revision: D2187844