Increase the number of retries
Summary: Several people have complained about `hh_client` giving up
prematurely, when 90% of the time it works on a subsequent invocation.
Bump up the number of retries and see if it makes people happier. Also,
now that we log the time it takes to connect + the frequency that we
show "giving up!", we'll be able to better tune this parameter if
necessary.

Reviewed By: @dlreeves

Differential Revision: D2194972