Port strtr from zend 5.6.10
strtr() invoked with an array $replace_pairs as the second argument appears to
perform much worse under HHVM than PHP5.

PHP's implementation of multiple-string search was formerly quite naïve, and it
was the subject of PHP bug 63893. In 2013, it was rewritten to utilize a clever
and efficient algorithm for multi-pattern searching (Wu-Manber), and then
further optimized. It looks like HHVM's implementation is based on the naive
implementation PHP sported before PHP 5.4, which explains the poor performance.

This patch ports over the strtr implementation in php 5.6.10. This implementation
prefers the existing strtr_fast algorithm and falls back to wu-manber when there
are too many patterns.