Move expansion of Taccess to Typing_phase.localize
Summary: When I first wrote the expansion logic for type access I didn't really
understand how Tvars and Tunresolved work, so at the time I would defer
expansion as late as possible in case the Tvars might change. Now that I
understand it better, when we expand a type access it should be done whenever
we localize the type.

This diff depends on D2177310. I move ##Taccess## to be a ##decl ty##, and use
the ##Typing_defs.dependent_type## as it's representation as a ##locl ty##.

This eliminates bugs due to forgetting to expand type constants, as well as
some cases that would be broken after D2177310. This also fixes a bug where
aliasing a type constant did not work properly.

The error messages have changed a bit, but I'll look into improving them after
all the other diffs impacting type constants land.

Reviewed By: @elgenie

Differential Revision: D2199960