Fix unwindResume for ASAN
Summary: ASAN generates red-zones for stack based objects, as well as for
each stack frame. It does this by storing 4 shadow bytes to mark
each red-zone when the function is entered. On exit, or on an exception,
it clears the shadow bytes.

It doesn't know about _Unwind_Resume, so when we called unwindResumeHelper,
it fills in the red-zone bytes, but never clears them, so that later code
can randomly die due to the bogus leftover bytes.

I tried marking it with FOLLY_DISABLE_ADDRESS_SANITIZER, and marking it
no-return, marking _Unwind_Resume no-return, and putting a not_reached()
after the call to _Unwind_Resume. But it always insisted on setting
the shadow bytes, and never managed to clear them before the call to
_Unwind_Resume.

So I modified the catch helper stub to mark tl_regState clean, so it can
just call _Unwind_Resume directly, and side-step the problem altogether.

Reviewed By: @swtaarrs

Differential Revision: D2202062