Remove LocalStateHook
Summary: Now that the original new refcount-opts is gone, there's only one
consumer of this information. Pull it back into FrameStateMgr.

Document and stop apologizing for the rule that minstr helpers must take
Local/Stack addresses as immediate srcs. We'll fix it if we fix it, but it's the
way things are for now.

Reviewed By: @jdelong

Differential Revision: D2203824