Track predicted types for individual SSATmps
Summary: This is necessary to preserve the predicted types of values as they
move between locals/stack slots. Before this diff, many tracelets would consist
of a just single bytecode like CGetL, because the CGetL only guarded on
UncountedInit and didn't keep a better prediction in the EvalStack slot that
held the loaded value. Now, the predictions are kept up-to-date and tracelets
are generally as long as without guard constriction on. The new SSATmp-based
predictions replace the predictions that were being kept in EvalStack, so those
are gone.

Reviewed By: @ottoni

Differential Revision: D2204388