Fix shutdown SEGV for uncounted APCCollections inside of APCArrays
Summary: This code was assuming APCLocalArray::sweep can't use g_context,
because g_context is already shut down when we run resetAllocator.  Before the
APCCollection diff, it actually couldn't use g_context, because we never make
uncounted arrays or strings inside of an APCArray.  (But we probably should try
that ...)

Reviewed By: @markw65

Differential Revision: D2206068