Add in a facility to pool curl handles between requests.
Summary: This diff introduces a new function, 'curl_init_pooled', which
pulls a curl handle from a persistent process-wide pool of curl
handles instead of making a new one (as 'curl_init' does). The
size of the process-wide pool and the maximum number of times a
curl handle will be reused in PHP are configurable via ini values.
By default, the connection pool will be turned off and a warning
will be issued if a user tries to use the facility without turning
it on.

This facility also introduces a new failure mode that one should
be aware of. When attempting to get a pooled curl handle but none
are available, the request will wait for an ini-value-configurable
number of seconds before timing out and fatal-ing. This serves as
a guard against resource starvation and will allow users to better
manager the curl handle pool.
Closes https://github.com/facebook/hhvm/pull/5509

Reviewed By: @JoelMarcey

Differential Revision: D2171830