Don't phi SpillFrames
Summary: We hit a case where an assertion fired due
to a temporary being kept live across a call. Thats
not really a problem, but in this case, it had created
a phi between an Object and a Cls (the ctx fields of
the respective SpillFrames). We currently don't handle
that correctly in code gen - although again, in this case
it didn't matter because the SpillFrame was dead at the
point it was inserted (which points to suboptimality in
the alias analysis code).

This diff adds an always_assert to catch the phi issue,
and also prevents phi-ing of SpillFrames until we resolve
the remaining issues.

Reviewed By: @ottoni

Differential Revision: D2218534