Send TE:chunked on HEAD, and double-check m_sendEnded
As requested in code review, send Transfer-Encoding:chunked on a HEAD
request, following the precedent of "at least one other big website".

Check m_sendEnded in onSendEnd(), and do nothing if it is set, to avoid
hitting the m_sendEnded case in ProxygenTransport::sendImpl, which a
comment indicates is not allowed. Hitting this case was possible for a
HEAD request with compression enabled. Before this change, it was also
possible to enter ProxygenTransport::onSendEndImpl() twice, in the case
of a non-compressed HEAD request. That is also fixed by this change.