Shapes::removeKey()
Summary: Continuing {D2130698} and {D2116812}, another thing that you might want to do with shapes, but currently can't. It uses the same logic to check that argument is a shape and that field might exists so I didn't repeat test cases from previous two diffs and focused on unique functionality of this function.

Reviewed By: @dlreeves

Differential Revision: D2130758