Remove the NOUSER define when compiling with MSVC, Cygwin, or MinGW.
This define causes nothing but problems and isn't actually needed. It causes the windows header to exclude a specific header, and, in order to use that header, we have to make sure to `#undef NOUSER` before we include `windows.h`, but that may be included by other headers with no knowledge of our needs, causing all sorts of messes. If you don't know the define is there, it's also a massive pain to figure out why your code isn't compiling (see: 4 hours digging through the internals of the windows headers to no avail).
Overall, it's better to just remove the define and save the headache.