Add a HHVM_REMOVE_MATCHES_FROM_LISTS CMake function.
There are numerous places in the source tree (mostly in the third-party repo) that loop trhough the files returned by `auto_sources` to remove files matching certain patterns, usually tests. The single loop on it's own is fine, but, when you start adding header files to it, as is needed to be useful in Visual Studio, it starts to become unmaintainable. To counter that, HHVM_REMOVE_MATCHES_FROM_LISTS is a single function that allows you to pass in the names of the lists, the matches to remove, and the matches to ignore.
This also replaces a couple of places using a basic foreach loop with a call to this function instead.