Port neo to MSVC.
All changes are pretty straight forward #ifdefs to include the correct headers, or implement it for windows except for the change in the return type of reg_search in the neo_str header. It is defined in the C file with a return value of BOOL, and, on posix, a char is the size of a bool, so this was never noticed. On windows however, a BOOL is not the size of a char, causing errors, which is why I noticed this at all. That is the only externally-facing API change.
It's worth noting though that that function is never actually called.