Make ArrayData ref-count start at 1
Summary: Similar to how ObjectData and StringData worked, the initial ref-count for
ArrayData started at zero in some cases, but was always eventually incremented
to one before being used. This is inefficient, so instead have the ref-count for
ArrayData start at one, and remove the now redundant inc-refs.

This has a much lesser impact than the previous conversions, as the hot paths
for arrays were already converted to having a starting ref-count of one. This
patch converts all the remaining cases.

Reviewed By: @binliu19

Differential Revision: D2199932