EHEntComp should implement a "strict weak ordering"
Summary: Apparently, comparators passed to `std::sort` are supposed to
implement a "strict weak ordering", that is, behave like `<` does. In
particular, they are supposed to return `false` when comparing equal
elements. In practice, the `std::sort` on Linux doesn't seem to ever
actually pass equal elements into here, so it hasn't mattered. However,
Clang on OS X //does// occasionally do this, and gives an incorrect
ordering if this rule isn't followed (thankfully, it trips an assert a
couple of lines below, instead of something more sinister and hard to
debug). Explicitly checking for equal elements and making sure we return
`false` fixes the problem.

Reviewed By: @jdelong

Differential Revision: D2183853