Remove uses of magicCall local where it's not possible to be true
Summary: This is a pretty small diff, but it's worth being separated out
on its own. After a particular point in fpushObjMethodWithBaseClass,
it's no longer possible for magicCall to be true. My guess is that it
was used as a result of copy-paste coding. So let's stop using it to
make it easier to continue refactoring.

Reviewed By: @edwinsmith

Differential Revision: D2144320