Only Create Expression Dependent Types When Necessary
Summary: After {D2199960} we can create unnecessary expression dependent types. In particular if we expanded a type constant into an expression dependent type, then any methods invoked on this type should not need any additional expression dependent types. As an example,

```
abstract class C {
  abstract const type T as C;

  abstract public function get(): this::T;

  public function test(C $c): void {
    $c_this = $this->get(); // <static>::T
    $c_this_2 = $c_this->get(); // should still be <static>::T
  }
}
```

Right now ##$c_this## and ##$c_this_2## would each be different expression dependent types, but after this diff they will be the same. To accomplish this when creating a new expression dependent type, we check if it is eligible to promoted to an expression dependent type. If the type is already expression dependent then we elect not create a new expression dependent type.

In order for any of this to be safe we require that the ##this::T## as a return type is resolved to the expression dependent type ##<static>::T## (i.e. 'this' is exactly the late bound class). This is more restrictive than what we currently require, but will still be useful after the next string of diffs that better tracks expression dependent types

Reviewed By: @jwatzman

Differential Revision: D2203422