Fix some leaks, and a bit more cleanup
Summary: We had a static array of all the SymbolTables, which
in turn had shared_ptrs to a lot of Expressions, which kept a
lot of things live after we thought we had freed them. It was no
longer used, so I killed it. Similarly, the CodeError table was
keeping Constructs alive. Between those we had about 10G of memory
live at the point that we started hhbbc which didn't need to be.

In addition, we were leaking anything allocated by the MemoryManager
for parser and emitter threads (meaning at least 2Mb per thread, and
sometimes much more).

I also killed some dead code that I came across while debugging
the leaks.

Reviewed By: @edwinsmith

Differential Revision: D2229543