Exceptions: Change trace handling
Summary: In some cases the setup of traces in Exception::initTrace() leads to unexpected behaviour. The properties for file and line won't be populated and sometimes the stack contains stack frames from systemlib.

An example can be seen here: http://3v4l.org/OTQQa (based on the test in slow/reflection/scalar_param_types/php-noforcehh.php)

However the code assumes that systemlib stack frames never provide file and/or line which might not be true all the time.

In a discussion in IRC about this issue @paulbiss suggested other approaches to solve that more reliable.

> [...] It might make sense to skip over native frames for consistency (though maybe native functions implemented in PHP should still be present in the trace...)
>
> maybe adding a flag for debug_backtrace to ignore native frames and setting that in Exception::$traceOpts would make sense
>
> [...] we'd probably want to do something with either debug_backtrace or by exposing some information about if a frame is builtin in the backtrace itself

Closes https://github.com/facebook/hhvm/pull/5580

Reviewed By: @sgolemon

Differential Revision: D2216854