Properly handle Zend compat resource lifetimes
Summary: The lifetimes of Zend compat resources are not currently handled
correctly. The Zend resources are referenced directly from a ResourceData
instance, but the Zend resource can be released manually, resulting in the
ResourceData pointing to a freed resource. Furthermore, the ResourceData's
ref-counts are not incremented properly.

This can manifest it in a crash, for instance, by calling ftp_connect(),
ftp_close(), then calling var_dump() on the resource returned by the connect.

This can be fixed by adopting an approach closer to Zend's approach. Instead of
a ResourceData referencing the Zend resource directly, instead just store an
index into the table of Zend resources. When the Zend resource is destroyed, the
entry in the table is released. However, since the ResourceData instances just
store the index, and the index is never re-used during a request, the
ResourceData can easily tell if the Zend resource is destroyed or not.

Reviewed By: @​bnell

Differential Revision: D2225681