Properly handle 'this' type for methods
Summary: The 'this' type acts a bit strange when invoking a method. If the class_id is CIparent or CIself, the type of 'this' will still be CIstatic. This is because the class_id is only used look up the method, but the context of evaluation is the current context. This means that the late bound type when executing ##self::foo()## or ##parent::foo()## is the same as ##static::foo()##.

This will pop up a couple errors since we are applying this to all ##this## types for a method and not just ##this::T##. This will be a temporary intermediate state until all the 'this' type unification diffs land.

Reviewed By: @jwatzman

Differential Revision: D2204325