Intersect local env when falling through a switch-case
Summary: In a switch statement with a fallthrough we would use the same environemnt
for the next case we encounter. However this is incorrect. Consider the
following case.

```
$x = 0; // int
switch (0) {
  case 1:
    $x = '';
    // FALLTHROUGH
  case 2:
    $x; // <-- #1
}

$x; // <-- #2
```

Currently at position #1 we will think $x is a string, since that is what we
assigned $x to be in the above case. However we could have switched directly
into case 2 without a fallthrough at which point $x would be an int.

This bug also impacts the result outside the switch statement. Because we
believe $x is of type string in both branches we believe $x is a string at
position #2 as well.

The solution is to intersect the local environment before and after type
checking the block in case 1. This intersected local environment is what should
be passed as the environment for case 2.

Reviewed By: @jwatzman, @int3

Differential Revision: D2230208