Eliminate a naming conflict
In `hphp_runtime_static`, the `HPHP::PCRECache::Accessor::Kind` enum has a value by the name of `Accessor`. This is the same as the name of an outer class and MSVC isn't happy about it.
This renames the `Accessor` value to `AccessorKind`. The enum is only used internally in the `Accessor` structure, so this shouldn't break anything.