Create stable expression dependent types for local variables
Summary: One final piece of analysis I want to add for tracking late bound types is for
locals. The goal of the analysis is:

1) Expression dependent types generated from a local variable should be stable
if the local has not be assigned.

2) After assigning local ##$foo## to local ##$bar##, ##$bar## should generate
the same dependent type as ##$foo##

The easiest way to achieve this was to add an additional value we track with
locals called an 'expression id'. An expression id is generated each time a
local variable is assigned a new value. If it is assigned a local variable,
then the expression id is transferred from one local to another, otherwise a
new expression id is generated.

I've written out in great detail what happens in the test cases I've added

Reviewed By: @jwatzman

Differential Revision: D2212586