Start of an experimental hack-based bytecode emitter
Summary: This is my WIP hack-based bytecode emitter for strict mode hack. It supports most basic statement and expression forms as well as simple uses of classes.

It is not particularly well integrated into anything at this point; it adds another option to hh_single_typecheck and emits hhas assembly that can be executed by hhvm with -vEval.AllowHhas=1.

Reviewed By: @int3

Differential Revision: D2184465