Eliminate a naming conflict for MSVC
Summary: In `hphp_runtime_static`, the `HPHP::PCRECache::Accessor::Kind` enum has a value by the name of `Accessor`. This is the same as the name of an outer class and MSVC isn't happy about it.
This renames the `Accessor` value to `AccessorKind`. The enum is only used internally in the `Accessor` structure, so this shouldn't break anything.
Closes https://github.com/facebook/hhvm/pull/5622

Reviewed By: @edwinsmith

Differential Revision: D2233846