Remove some includes of util/deprecated/base.h
Summary: Made a lot of foo -> std::foo in compiler/option.cpp to match option.h.
Had to add some includes to util/deprecated/declare-boost-types.h to make stuff
compile but I think it's worth it.

The remaining includes are in compiler/hphp.h, and in system/idl/idl.php.  Those
are going to be difficult, so I'll leave them for a future diff.

Reviewed By: @markw65

Differential Revision: D2232182