Remove restriction on the 'this' type
Summary: Now that all the machinery for making 'this' expression dependent is in place we can lift the restriction on allowing its usage for only a hardcoded list of types. I went through and created several test cases to convince myself the analysis works like we expect. There are only two cases where we still restrict using the 'this' type.

* It is forbidden as a type for static class variables. The storage of class variables is weird since it can be shared for a whole class hierarchy.
* It is forbidden as a constraint on a class type parameter. This is a recursive type, but doesn't cause infitine loops in anyway. It is just strange so its best not to allow it.

Reviewed By: @int3

Differential Revision: D2235191