Port neo to MSVC
Summary: Most of the changes are pretty straight forward #ifdefs to include the correct headers, or implement it for windows.

The changes to the casts in the calls to `uListGet` in `neo_err.c` are because MSVC, correctly, complains about the argument types not matching.

There is also a change in the return type of reg_search in `neo_str.h`. It is defined in the C file with a return value of BOOL, and, on posix, a char is the size of a bool, so this was never noticed. On windows however, a BOOL is not the size of a char, causing errors, which is why I noticed this at all. That is the only externally-facing API change. It's worth noting though that that function is never actually called in HHVM or neo itself.
Closes https://github.com/facebook/hhvm/pull/5593

Reviewed By: @

Differential Revision: D2222104

Pulled By: @sgolemon