Call gen-ext-hhvm directly rather than through a bash script
The bash script doesn't really do anything but re-order the arguments, and, because you can't set environment variables when calling a bash script from CMake on Windows, the way it was being done didn't work.
This calls the gen-ext-hhvm target directly with the arguments in the expected order.

Assuming the FB internal build calls the bash script `gen-ext-hhvm.sh`, it will have to be adjusted to call `gen-ext-hhvm` directly as well.