Make NEVER_INLINE usages MSVC compatible
Because MSVC doesn't like `__declspec()` attributes after the arguments.
Hopefully I got all the uses in this PR, but I didn't check every single use of the attribute, as the vast majority are already placed before the function name.

This will conflict with #5632 (D41565) due to a spot in the asio extension where both `ATTRIBUTE_NORETURN` and `NEVER_INLINE` attributes needed to be moved.