Have roundImpl explicitly accept a function pointer type
Because MSVC overloads `ceil` and `floor` with `float`, `double`, and `long double` forms, we have to force it to select only one of those forms.
Unfortunately, there is not a specific name for the `double` overload, otherwise I'd just `#ifdef _MSC_VER` and use that other name.

If this was done as a template for speed reasons, then the same effect can be achieved with a force inline attribute.