Call gen-ext-hhvm directly rather than through a bash script
Summary: The bash script doesn't really do anything but re-order the arguments, and, because you can't set environment variables when calling a bash script from CMake on Windows, the way it was being done didn't work.
This calls the gen-ext-hhvm target directly with the arguments in the expected order.

Closes https://github.com/facebook/hhvm/pull/5668

Reviewed By: @

Differential Revision: D2235927

Pulled By: @sgolemon