Fix anonymous function unification
Summary: {D2106548} was not the right fix - even though spurious errors are caused by anonymous function unification resulting in Tunresolved[Tunresolved[...]], and flattening it solves the issue, it's not the right thing to do because it expands (and removes) type variables that might be relevant later.

The whole unification scheme for anonymous functions was incorrect anyway - we were always claiming that any two anonymous functions unify without errors. Saying that they never unify is less likely to introduce unsoundness, and doesn't break anything else.

Reviewed By: @jwatzman, @elgenie

Differential Revision: D2212522