Port the parser to MSVC
Summary: This switches a couple of macros from C99 forms to normal variadic forms.
This also changes a couple of locals from `auto` to `size_t` in order to convince MSVC to resolve to a single overload.
This also adds an include for `hphp/util/portability.h`, which will, in turn, include a few compatibility shims that are needed.
Closes https://github.com/facebook/hhvm/pull/5647

Reviewed By: @

Differential Revision: D2235805

Pulled By: @sgolemon