Allow CMake to choose the debug or release TBB libraries
Summary: This makes a difference for targets that support multiple configuration types, MSVC for instance. Currently the release libraries are unconditionally linked in, and this breaks Windows, because the TBB headers see that it's debug mode and try to force you to link in the debug version of the library (as-is only really possible on MSVC as far as I know). This still falls back to the previous variable pairs for targets not supporting multiple configuration types.
This also fixes the indentation of a few blocks at the end of the file that took me a second to figure out what it was doing.
Closes https://github.com/facebook/hhvm/pull/5661

Reviewed By: @sgolemon

Differential Revision: D2235913