Have roundImpl explicitly accept a function pointer type
Summary: Because MSVC overloads `ceil` and `floor` with `float`, `double`, and `long double` forms, we have to force it to select only one of those forms.
Unfortunately, there is not a specific name for the `double` overload, otherwise I'd just `#ifdef _MSC_VER` and use that other name.

If this was done as a template for speed reasons, then the same effect can be achieved with a force inline attribute.
Closes https://github.com/facebook/hhvm/pull/5684

Reviewed By: @

Differential Revision: D2237249

Pulled By: @sgolemon