Inline emitDecRef() so dec and compare operations are grouped together.
Summary: Upcoming work on refcounting will be simpler if we keep the code snippets
encapsulated such that code that sets status flags is nearby code that
uses the status flags.

emitDecRef() was only being called by other utility functions that
already do an adequate job of encapsulating refcount logic, so inline it
into the few callsites.

Reviewed By: @ricklavoie

Differential Revision: D2240923