Fix GCC 5.1.1 compilation errors
Summary: This fixes some of the compilation breaks resulting from warnings being promoted
to errors in GCC 5.1.1 vs GCC 4.9. It's now possible to obtain a release build
however the debug build still continues to break.

In addition to this third-party/re2 should be updated to to avoid narrowing
conversion errors.
Closes https://github.com/facebook/hhvm/pull/5472

Reviewed By: @sgolemon

Differential Revision: D2146527